+++ packages/core/src/core/layer.ts
@@ -1,7 +1,6 @@
 import { IdetikContext } from "../idetik";
 import { RenderableObject } from "./renderable_object";
 import { clamp } from "../utilities/clamp";
-import { ChannelProps } from "@/objects/textures/channel";
 
 export type LayerState = "initialized" | "loading" | "ready";
 export type blendMode = "normal" | "additive" | "subtractive" | "multiply";
@@ -107,13 +106,4 @@ export abstract class Layer {
   protected clearObjects() {
     this.objects_ = [];
   }
-}
-
-/** Layer that exposes channel controls. */
-export interface ChannelsEnabled {
-  channelProps: ChannelProps[] | undefined;
-  setChannelProps(channelProps: ChannelProps[]): void;
-  resetChannelProps(): void;
-  addChannelChangeCallback(callback: () => void): void;
-  removeChannelChangeCallback(callback: () => void): void;
 }
